refactor(hermes): split config generator modules
